refactor: padding 변수 타입 수정 및 코드 명확성 개선

+++ Engine/EngineCore/Source/Core/Memory/MemoryResource/TrackedMemoryResource.cpp
@@ -43,7 +43,7 @@ void* TrackedMemoryResource::do_allocate(size_t size, size_t align)
     MemoryAllocHeader* header = reinterpret_cast<MemoryAllocHeader*>(static_cast<uint8*>(user_ptr) - HEADER_SIZE);
 
     // 헤더가 실제 할당된 블럭으로부터 얼만큼 떨어져 있는지 계산
-    const uint32 padding = static_cast<uint8>(
+    const uint8 padding = static_cast<uint8>(
         reinterpret_cast<uintptr_t>(header) - reinterpret_cast<uintptr_t>(raw_block)
     );
